Improve CSS and responsiveness - Relates #2 #7 #13
* I adopted a mobile-first mindset, taking advantage of flexbox and
experimenting with upscaling the website units using a single media
query on the :root element.

* Converted the ‘toggle-button’ on small devices from a link to a
button element for improved semantics.

* Improved toggle button using JavaScript to change the value of
the [aria-expanded] CSS attribute and open/close the menu without
applying a custom ‘active’ class.

* Modified styling of photos in the work section, icons in the
services section and figure elements in the about section to
improve responsiveness.

* The custom font is now being applied; it wasn’t before.
Added font-display property on the body element with a value
of swap to render a fallback font until the custom font is ready.
Also added text shadow on H1 to ‘make it pop’ more.

* Added box-shadow transition on buttons.

* On second thought I removed the 100vh on the services section
to allow it to naturally stretch in height if we will add more
content, at the moment it would just look weird and empty, but
let me know if you agree or you’d rather leave it! 😄